# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Risk Control API",
        "description": "API para sistema de control de riesgo en operaciones de trading",
        "contact": {
            "email": "soporte@riskcontrol.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api/v1",
            "description": "Servidor local"
        }
    ],
    "paths": {
...
        "/rules": {
            "get": {
                "tags": [
                    "RiskRules"
                ],
                "summary": "Listar reglas de riesgo",
                "description": "Obtener lista de reglas",
                "operationId": "0e384fb0a0e8e5db0cb98a08f17d8324",
                "parameters": [
                    {
                        "name": "type",
                        "in": "query",
                        "description": "Filtrar por tipo de regla",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "DURATION",
                                "VOLUME",
                                "OPEN_TRADES"
                            ]
                        }
                    },
                    {
                        "name": "severity",
                        "in": "query",
                        "description": "Filtrar por severidad",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "HARD",
                                "SOFT"
                            ]
                        }
                    },
                    {
                        "name": "is_active",
                        "in": "query",
                        "description": "Filtrar por estado activo",
                        "schema": {
                            "type": "boolean"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Elementos por página",
                        "schema": {
                            "type": "integer",
                            "default": 15
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de reglas",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/PaginatedResponse"
                                        },
                                        {
                                            "properties": {
                                                "data": {
                                                    "type": "array",
                                                    "items": {
                                                        "$ref": "#/components/schemas/RiskRule"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthorized"
                    }
                },
                "security": [
                    {
                        "apiKey": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "RiskRules"
                ],
                "summary": "Crear regla de riesgo",
                "description": "Crear nueva regla",
                "operationId": "0a0eb4bd62f30e69031f7967335e3beb",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RiskRule"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Regla creada exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Risk rule created successfully"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/RiskRule"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthorized"
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationError"
                    }
                },
                "security": [
                    {
                        "apiKey": []
                    }
                ]
            }
        },
        "/rules/{id}": {
            "get": {
                "tags": [
                    "RiskRules"
                ],
                "summary": "Obtener regla por ID",
                "description": "Obtener regla por ID",
                "operationId": "71e63db26877f9797c047361a8a43f07",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la regla",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Regla obtenida",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/RiskRule"
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthorized"
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFound"
                    }
                },
                "security": [
                    {
                        "apiKey": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "RiskRules"
                ],
                "summary": "Actualizar regla",
                "description": "Actualizar regla",
                "operationId": "e08413db4c90fd204e4588df0ae9bd1f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la regla",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RiskRule"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Regla actualizada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Risk rule updated successfully"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/RiskRule"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthorized"
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFound"
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationError"
                    }
                },
                "security": [
                    {
                        "apiKey": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "RiskRules"
                ],
                "summary": "Eliminar regla",
                "description": "Eliminar regla",
                "operationId": "e9b42d2dada869694bfe7fd64f75892a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la regla",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Regla eliminada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Risk rule deleted successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthorized"
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFound"
                    }
                },
                "security": [
                    {
                        "apiKey": []
                    }
                ]
            }
        },
        "/rules/{id}/toggle-active": {
            "post": {
                "tags": [
                    "RiskRules"
                ],
                "summary": "Alternar estado activo",
                "description": "Alternar estado activo de regla",
                "operationId": "1d55eae0ff347a6d2bbac288cd92feda",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la regla",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Estado actualizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Risk rule status updated"
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "is_active": {
                                                    "type": "boolean",
                                                    "example": true
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthorized"
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFound"
                    }
                },
                "security": [
                    {
                        "apiKey": []
                    }
                ]
            }
        },
...
        "/rules/types/info": {
            "get": {
                "tags": [
                    "RiskRules"
                ],
                "summary": "Obtener información de tipos de reglas",
                "description": "Obtener información de tipos de reglas",
                "operationId": "cf7636a77a370a7e2cdff8a1e1710c19",
                "responses": {
                    "200": {
                        "description": "Información de tipos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "types": {
                                            "type": "object",
                                            "example": {
                                                "DURATION": "Duración de Trade",
                                                "VOLUME": "Consistencia de Volumen",
                                                "OPEN_TRADES": "Trades Abiertos"
                                            }
                                        },
                                        "types_info": {
                                            "description": "Información detallada por tipo",
                                            "type": "object"
                                        },
                                        "severities": {
                                            "type": "object",
                                            "example": {
                                                "HARD": "Regla Dura",
                                                "SOFT": "Regla Suave"
                                            }
                                        },
                                        "action_types": {
                                            "description": "Tipos de acciones disponibles",
                                            "type": "array",
                                            "items": {
                                                "type": "string",
                                                "example": "EMAIL"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthorized"
                    }
                },
                "security": [
                    {
                        "apiKey": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "PaginatedResponse": {
                "title": "Respuesta Paginada",
                "properties": {
                    "data": {
                        "type": "array",
                        "items": {
                            "type": "object"
                        }
                    },
                    "links": {
                        "type": "object"
                    },
                    "meta": {
                        "type": "object"
                    }
                },
                "type": "object"
            },
...
        },
        "responses": {
...
            "Unauthorized": {
                "description": "No autorizado",
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "error": {
                                    "type": "string",
                                    "example": "Unauthorized"
                                },
                                "message": {
                                    "type": "string",
                                    "example": "Clave API inválida"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
...
        },
        "securitySchemes": {
            "apiKey": {
                "type": "apiKey",
                "name": "X-API-KEY",
                "in": "header"
            }
        }
    },
...
}